code/requirements.txt
opencv-contrib-python == 4.9.0.80
opencv-python == 4.9.0.80
numpy == 1.26.3
pillow == 10.2.0